CMW-51/track_dao_settings
Fixed after review
